pdf_ocr/requirements.txt
fastapi
uvicorn
pdfplumber
pillow
numpy
opencv-python-headless
pytesseract
fuzzywuzzy
sqlalchemy